+++ CHANGELOG.md
@@ -21,6 +21,7 @@ and this project adheres to [Semantic Versioning][].
 -   Legend order is now deterministic (#143)
 -   Images no longer normalised by default (#150)
 -   Colorbar no longer autoscales to [0, 1] (#155)
+-   Plotting shapes after a spatial query is now possible (#163)
 
 ## [0.0.4] - 2023-08-11